[coordinates/clustering] changed dtype of dtrajs from int64 to int32

+++ pyemma/coordinates/tests/test_assign.py
@@ -145,7 +145,7 @@ class TestCluster(unittest.TestCase):
 
     def test_output_type(self):
         c = self.ass
-        assert c.output_type() == np.int64
+        assert c.output_type() == np.int32
 
     def test_parametrize(self):
         c = self.ass
